oli: move integration test to dir tests
