+++ oli/src/commands/cli.rs
@@ -61,36 +61,3 @@ fn build_object(path: &str) -> Result<Object> {
     let operator = super::profile::build_operator(&cp_path)?;
     Ok(operator.object(&cp_path.path()))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use assert_cmd::prelude::*;
-    
-    use std::env;
-    use std::fs;
-    use std::path::Path;
-    use std::process::Command;
-
-    #[tokio::test]
-    async fn test_basic_cp() -> Result<()> {
-        let dir = env::temp_dir();
-        fs::create_dir_all(dir.clone())?;
-        let src_path = Path::new(&dir).join("src.txt");
-        let dst_path = Path::new(&dir).join("dst.txt");
-        let expect = "hello";
-        fs::write(&src_path, expect)?;
-
-        let mut cmd = Command::cargo_bin("oli")?;
-
-        cmd.arg("cp")
-            .arg(src_path.as_os_str())
-            .arg(dst_path.as_os_str());
-        cmd.assert().success();
-
-        let actual = fs::read_to_string(&dst_path)?;
-        assert_eq!(expect, actual);
-        Ok(())
-    }
-}
